Delete id type with dry run

+++ src/api/NomaNova.Ojeda.Services/AssetIdTypes/AssetIdTypesService.cs
@@ -1,24 +1,31 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
 using NomaNova.Ojeda.Core.Domain.AssetIdTypes;
+using NomaNova.Ojeda.Core.Domain.AssetTypes;
 using NomaNova.Ojeda.Core.Exceptions;
 using NomaNova.Ojeda.Data.Repositories;
 using NomaNova.Ojeda.Models.Dtos.AssetIdTypes;
 using NomaNova.Ojeda.Models.Shared;
 using NomaNova.Ojeda.Services.AssetIdTypes.Interfaces;
 using NomaNova.Ojeda.Services.AssetIdTypes.Validators;
+using NomaNova.Ojeda.Utils.Extensions;
 
 namespace NomaNova.Ojeda.Services.AssetIdTypes;
 
 public class AssetIdTypesService : BaseService<AssetIdType>, IAssetIdTypesService
 {
+    private readonly IRepository<AssetType> _assetTypesRepository;
+
     public AssetIdTypesService(
         IMapper mapper,
-        IRepository<AssetIdType> assetIdTypesRepository) : base(mapper, assetIdTypesRepository)
+        IRepository<AssetIdType> assetIdTypesRepository,
+        IRepository<AssetType> assetTypesRepository) : base(mapper, assetIdTypesRepository)
     {
+        _assetTypesRepository = assetTypesRepository;
     }
 
     public async Task<AssetIdTypeDto> GetByIdAsync(string id, CancellationToken cancellationToken = default)
@@ -75,7 +82,8 @@ public class AssetIdTypesService : BaseService<AssetIdType>, IAssetIdTypesServic
         return Mapper.Map<AssetIdTypeDto>(assetIdType);
     }
 
-    public async Task DeleteAsync(string id, CancellationToken cancellationToken = default)
+    public async Task<DryRunDeleteAssetIdTypeDto> DeleteAsync(string id, bool dryRun,
+        CancellationToken cancellationToken = default)
     {
         var assetIdType = await Repository.GetByIdAsync(id, cancellationToken);
 
@@ -83,7 +91,30 @@ public class AssetIdTypesService : BaseService<AssetIdType>, IAssetIdTypesServic
         {
             throw new NotFoundException();
         }
-        
+
+        var deleteAssetIdTypeDto = new DryRunDeleteAssetIdTypeDto();
+
+        // Fetch impacted asset types
+        var assetTypes = await _assetTypesRepository.GetAllAsync(query =>
+        {
+            return query.Where(_ => _.AssetIdTypeId == id);
+        }, cancellationToken);
+
+        deleteAssetIdTypeDto.AssetTypes = assetTypes.Select(_ => Mapper.Map<NamedEntityDto>(_)).ToList();
+
+        if (dryRun)
+        {
+            return deleteAssetIdTypeDto;
+        }
+
+        // Do not proceed when asset types are still linked to this id type
+        if (assetTypes.HasElements())
+        {
+            throw new ValidationException(
+                nameof(AssetIdType.Id), "Cannot delete item as it is in use.");
+        }
+
         await Repository.DeleteAsync(assetIdType, cancellationToken);
+        return deleteAssetIdTypeDto;
     }
 }
